21 - Enhance error logging and refactor notes retrieval (#23)
* Enhance error logging and refactor notes retrieval

Introduced a logError function for use in promise catches to improve error logging. The way notes are retrieved and set has been refactored with a new function, setNoteContentIfDataExists, for cleaner code and improved maintainability. The note handling mechanism within the service_worker.js file has also been streamlined.

* Fix saved notes bug

This commit refactors the setNoteContentIfDataExists function to include removal of notes when data is empty and error handling. It introduces the removeNotes function and misc changes to streamline the note handling mechanism.