Set up environment variables for manifest file
This commit sets up the infrastructure to use environment variables in the manifest file for the chrome extension. Manifest.json is renamed to manifest.template.json, and a new buildEnv script is introduced to generate the manifest.json file using the environment variables. A dotenv dependency is added to manage these environment variables.