14 openai proxy (#45)
* Set up environment variables for manifest file

This commit sets up the infrastructure to use environment variables in the manifest file for the chrome extension. Manifest.json is renamed to manifest.template.json, and a new buildEnv script is introduced to generate the manifest.json file using the environment variables. A dotenv dependency is added to manage these environment variables.

* Add auth button and implement OAuth process

An authentication button has been added to the user interface. The OAuth authentication method was set up to fetch the user's email and use it for back-end communications. OpenAI's call method has been adjusted to fall back to an alternative proxy-based method if the default authentication fails.

* Update manifest template and auth checking

Use environment variables for the versioning in the manifest template. Renamed the 'buildEnv.js' file to 'buildManifest.js' and updated this in the npm script. An auth checking function was added to the 'popup.js' file to change the Auth button depending on context.

* Use instance ID for tracking API Requests

* Remove identity permission requirement

* This commit eliminates the need for authentication via Google OAuth. In addition, it includes the integration of Sentry's JavaScript browser tracking.

* Add source maps for Sentry

* Fix bug where openai was still being called even when proxy should be used

* Update test setup and event listeners

Refactored the testing setup to initialize the mock `global.chrome` object in a `beforeAll` block. Added jest to the `setupFiles` in `package.json`.

* Update URL following token revocation, add proxy URL to host permissions