44 cycle time (#53)
* Add cycle time calculation to content script

A new function, setCycleTime() was created which calculates and displays the cycle time for stories in the 'Completed' state. A utility function, isInState(), from the developmentTime.js script was also updated to be an exportable function.

* Refactor `developmentTime.js` and update `contentScript.js` for cycle time calculation

Several functions in `developmentTime.js` were refactored to be exportable and to improve calculation of development times. The `contentScript.js` script was updated to use these changes to calculate the cycle time for completed stories.

* Use Sentry release versioning

* Remove unused import