Add ability to customise workflow states
This commit adds the ability for users to customise the names of their workflow states in the "Shortcut Assistant" application's settings. This functionality was achieved by including new input fields in the settings UI and integrating corresponding data handlers in the back-end scripts. In addition, to facilitate this change, various utilities and test files were modified and updated, correspondingly.