Add functionality for saving Shortcut API Token
This update introduces functionality to save Shortcut API tokens in the pop-up. It also includes refactoring of the save button functionality for more generalized use. Additionally, it introduces the functionality to register users with user details and the Shortcut token. Environment handling was revamped to use a .local.env file during development.